Add computing Fibonacci number function in "test" operation.
When type "test f 10" for example in the qemu, the result of Fibonacci number at 10 will be shown up in the terminal.
With this fixed input format, the program will recognize the character 'f' and then get the folling number as the computing parameter.
After the computing function finished, the result will be shown up by the function "fio_printf".