Add checking function that whether a number is a prime or not.
With fixed format "test p 100" for example, the function will check the input number is a prime or not.
If the input number is a prime, the remainder of the division of the input number and the number which is smaller than the sqrt of the input number is zero.